+++ app/charts/column/columns-stacked-state.tsx
@@ -215,7 +215,7 @@ const useColumnsStackedState = (
     }
 
     return group(chartData, getX);
-  }, [chartData, getX, sumsByX, getY, yMeasure.id, normalize]);
+  }, [normalize, chartData, getX, yMeasure.id, getY, sumsByX]);
 
   const chartWideData = useMemo(() => {
     return getWideData({