+++ app/configurator/components/chart-options-selector/imputation-field.tsx
@@ -1,0 +1,96 @@
+import { t, Trans } from "@lingui/macro";
+import get from "lodash/get";
+import { useCallback } from "react";
+
+import { Select } from "@/components/form";
+import { ChartConfig, IMPUTATION_TYPES, ImputationType } from "@/config-types";
+import {
+  ControlSection,
+  ControlSectionContent,
+  SectionTitle,
+} from "@/configurator/components/chart-controls/section";
+import { getFieldLabel } from "@/configurator/components/field-i18n";
+import {
+  isConfiguring,
+  useConfiguratorState,
+} from "@/configurator/configurator-state";
+
+export const ImputationField = ({
+  chartConfig,
+}: {
+  chartConfig: ChartConfig;
+}) => {
+  const [, dispatch] = useConfiguratorState(isConfiguring);
+  const getImputationTypeLabel = (type: ImputationType) => {
+    switch (type) {
+      case "none":
+        return t({
+          id: "controls.imputation.type.none",
+          message: "-",
+        });
+      case "zeros":
+        return t({
+          id: "controls.imputation.type.zeros",
+          message: "Zeros",
+        });
+      case "linear":
+        return t({
+          id: "controls.imputation.type.linear",
+          message: "Linear interpolation",
+        });
+      default:
+        const _exhaustiveCheck: never = type;
+        return _exhaustiveCheck;
+    }
+  };
+  const updateImputationType = useCallback<(type: ImputationType) => void>(
+    (type) => {
+      dispatch({
+        type: "IMPUTATION_TYPE_CHANGED",
+        value: {
+          type,
+        },
+      });
+    },
+    [dispatch]
+  );
+
+  const imputationType: ImputationType = get(
+    chartConfig,
+    ["fields", "y", "imputationType"],
+    "none"
+  );
+
+  return (
+    <ControlSection collapse>
+      <SectionTitle
+        iconName="infoCircle"
+        warnMessage={
+          imputationType === "none"
+            ? t({
+                id: "controls.section.imputation.explanation",
+                message:
+                  "For this chart type, replacement values should be assigned to missing values. Decide on the imputation logic or switch to another chart type.",
+              })
+            : undefined
+        }
+      >
+        <Trans id="controls.section.imputation">Missing values</Trans>
+      </SectionTitle>
+      <ControlSectionContent component="fieldset" gap="none">
+        <Select
+          id="imputation-type"
+          label={getFieldLabel("imputation")}
+          options={IMPUTATION_TYPES.map((d) => ({
+            value: d,
+            label: getImputationTypeLabel(d),
+          }))}
+          value={imputationType}
+          onChange={(e) => {
+            updateImputationType(e.target.value as ImputationType);
+          }}
+        />
+      </ControlSectionContent>
+    </ControlSection>
+  );
+};